remove 'src' directory simplify the command line runner
