+++ .gitignore
@@ -1,4 +1,6 @@
 *.ipynb
 *.pyc
 *.csv
-/tests/.pytest_cache/
+.pytest_cache/
+venv/
+.idea/
