+++ tests/test_objects.py
@@ -1,9 +1,9 @@
 from typing import List, Type
 
 import pytest
-from _pytest.compat import nullcontext
+from contextlib import nullcontext as DoesNotRaise
 
-from src.config import Body, Orbit
+from utils.objects import Body, Orbit
 
 
 @pytest.mark.parametrize(
@@ -16,14 +16,14 @@ from src.config import Body, Orbit
                 Body(x=1, vx=1, y=1, vy=1, mass=1, name="body_2"),
             ],
             10,
-            None,
+            DoesNotRaise(),
         ),
-        ("INVALID_ORBIT", [], 10, ValueError),
+        ("INVALID_ORBIT", [], 10, pytest.raises(ValueError)),
         (
             "INVALID_ORBIT",
             [Body(x=0, vx=0, y=0, vy=0, mass=1, name="body_1")],
             10,
-            ValueError,
+            pytest.raises(ValueError),
         ),
     ],
 )
@@ -31,5 +31,5 @@ def test_orbit_constructor(
     name: str, bodies: List[Body], t: int, exception: Type[BaseException]
 ):
     """ Tests that Orbit constructor raises ValueError if not enough bodies supplied """
-    with pytest.raises(exception) if exception else nullcontext():
+    with exception:
         Orbit(name=name, bodies=bodies, t=t)
